# Implement SWAR indexOf byte search