@franz1981 2020-10-28
@hyperxpro 
> Any benchmark results for this?

That's a draft PR :)
Still working on the logic so doesn't make sense to bench it yet