@hyperxpro 2020-10-28
```
13:46:56 [ERROR] /code/buffer/src/main/java/io/netty/buffer/ByteBufUtil.java:442: Line is longer than 120 characters (found 128). [LineLength]
13:46:56 [ERROR] /code/buffer/src/test/java/io/netty/buffer/AbstractByteBufTest.java:2155: two or more consecutive empty lines [RegexpMultiline]
```