@richardstartin 2020-10-28
I'm really interested to find out if this works well in netty. I think there are substring searches to look at too, which should benefit from the same wider access.