@franz1981 2020-10-28
> I'm really interested to find out if this works well in netty. I think there are substring searches to look at too, which should benefit from the same wider access.

Me too! And curious to check if the `TODO` I've added can make any difference too!

I'll come up with a bench soon, probably using yours on the article as a base: many thanks again for both the article and your help here bud :)