@franz1981 2020-10-28
@richardstartin @hyperxpro 

With a very unfair data set length and no optimization at all (including the right advice of @richardstartin re creating separate versions LE/BE of a method), I'm getting:
```
Benchmark                              (logPermutations)  (needleByte)  (seed)  (size)   Mode  Cnt    Score   Error   Units
ByteBufIndexOfBenchmark.bufferIndexOf                 16             0       1       8  thrpt   16   41.017 ± 0.418  ops/us
ByteBufIndexOfBenchmark.bufferIndexOf                 16             0       1      16  thrpt   16   35.957 ± 0.556  ops/us
ByteBufIndexOfBenchmark.bufferIndexOf                 16             0       1      32  thrpt   16   26.767 ± 0.450  ops/us
ByteBufIndexOfBenchmark.bufferIndexOf                 16             0       1      64  thrpt   16   17.747 ± 0.403  ops/us
ByteBufIndexOfBenchmark.swarIndexOf                   16             0       1       8  thrpt   16  100.340 ± 3.007  ops/us
ByteBufIndexOfBenchmark.swarIndexOf                   16             0       1      16  thrpt   16   79.562 ± 2.257  ops/us
ByteBufIndexOfBenchmark.swarIndexOf                   16             0       1      32  thrpt   16   60.568 ± 0.990  ops/us
ByteBufIndexOfBenchmark.swarIndexOf                   16             0       1      64  thrpt   16   37.069 ± 0.776  ops/us
```
It seems promising :) @normanmaurer  wdyt?