@richardstartin 2020-10-28
@franz1981 it's probably worth testing with non random data too because it might often be quite predictable. I'm not really familiar with what this is used for in netty.