@richardstartin 2020-10-28
@franz1981 so faster even when predictable? Nice!