@franz1981 2020-10-28
@richardstartin predicatable for the machine, not sure for the JIT eheh :P
But yes, it still seems a win-win, let me check what happen with (N*8) + M with M < 8 sized data too!