@franz1981 2020-10-28
And some results from mixed workloads with predictable data:
```
Benchmark                              (logPermutations)  (needleByte)  (seed)  (size)   Mode  Cnt    Score    Error   Units
ByteBufIndexOfBenchmark.bufferIndexOf                  4             0       1       7  thrpt   16  103.167 ±  3.625  ops/us
ByteBufIndexOfBenchmark.bufferIndexOf                  4             0       1       8  thrpt   16  102.151 ±  3.264  ops/us
ByteBufIndexOfBenchmark.bufferIndexOf                  4             0       1      11  thrpt   16   95.552 ±  0.905  ops/us
ByteBufIndexOfBenchmark.bufferIndexOf                  4             0       1      15  thrpt   16   77.921 ±  0.147  ops/us
ByteBufIndexOfBenchmark.swarIndexOf                    4             0       1       7  thrpt   16  105.330 ±  0.664  ops/us
ByteBufIndexOfBenchmark.swarIndexOf                    4             0       1       8  thrpt   16   97.496 ± 12.752  ops/us
ByteBufIndexOfBenchmark.swarIndexOf                    4             0       1      11  thrpt   16   93.064 ±  0.797  ops/us
ByteBufIndexOfBenchmark.swarIndexOf                    4             0       1      15  thrpt   16   72.519 ±  1.747  ops/us
```