@franz1981 2020-10-28
While with progressive unpredictability:
```
Benchmark                              (logPermutations)  (needleByte)  (seed)  (size)   Mode  Cnt   Score   Error   Units
ByteBufIndexOfBenchmark.bufferIndexOf                 11             0       1      10  thrpt   32  81.586 ± 2.863  ops/us
ByteBufIndexOfBenchmark.bufferIndexOf                 11             0       1      15  thrpt   32  62.978 ± 2.456  ops/us
ByteBufIndexOfBenchmark.bufferIndexOf                 15             0       1      10  thrpt   32  39.627 ± 0.507  ops/us
ByteBufIndexOfBenchmark.bufferIndexOf                 15             0       1      15  thrpt   32  36.405 ± 0.322  ops/us
ByteBufIndexOfBenchmark.swarIndexOf                   11             0       1      10  thrpt   32  87.960 ± 0.571  ops/us
ByteBufIndexOfBenchmark.swarIndexOf                   11             0       1      15  thrpt   32  70.150 ± 1.678  ops/us
ByteBufIndexOfBenchmark.swarIndexOf                   15             0       1      10  thrpt   32  84.940 ± 0.813  ops/us
ByteBufIndexOfBenchmark.swarIndexOf                   15             0       1      15  thrpt   32  67.781 ± 2.355  ops/us
```