@franz1981 2020-10-29
I see that 
```java
final boolean useLE = nativeOrder == ByteOrder.LITTLE_ENDIAN;
``` 
Is a const condition: I would like to check if the compiler is treating it correctly as constant