@franz1981 2020-10-31
@richardstartin 
I've found the reason why the SWAR version is competitive for small sizes and predictable cases too: loop unrolling!
using a counted loop sized with `& 7` seems to give the JIT the confidence to loop unroll it, getting good results. 
I haven't had time to try this with OpenJ9 (that doesn't play well with compiler directives) or Graal, but seems fun :)