@franz1981 2020-10-31
I've improved the benchmarks and I'm getting some bad results with `direct ByteBufs`, need some investigations:
```
Benchmark                                   (direct)  (logPermutations)  (needleByte)  (noUnsafe)  (seed)  (size)   Mode  Cnt   Score   Error   Units
ByteBufIndexOfBenchmark.bufferIndexOf          false                  4             0       false       1       7  thrpt   16  98.510 ± 2.296  ops/us
ByteBufIndexOfBenchmark.bufferIndexOf          false                  4             0       false       1      16  thrpt   16  69.948 ± 0.659  ops/us
ByteBufIndexOfBenchmark.bufferIndexOf           true                  4             0       false       1       7  thrpt   16  97.303 ± 0.844  ops/us
ByteBufIndexOfBenchmark.bufferIndexOf           true                  4             0       false       1      16  thrpt   16  76.216 ± 0.506  ops/us
ByteBufIndexOfBenchmark.swarIndexOf            false                  4             0       false       1       7  thrpt   16  92.764 ± 1.508  ops/us
ByteBufIndexOfBenchmark.swarIndexOf            false                  4             0       false       1      16  thrpt   16  81.231 ± 1.548  ops/us
ByteBufIndexOfBenchmark.swarIndexOf             true                  4             0       false       1       7  thrpt   16  82.591 ± 1.040  ops/us
ByteBufIndexOfBenchmark.swarIndexOf             true                  4             0       false       1      16  thrpt   16  77.056 ± 0.579  ops/us
```