@franz1981 2020-11-02
This is the performance with the very last version:
NEW:
```
Benchmark                             (direct)  (logPermutations)  (needleByte)  (noUnsafe)  (pooled)  (seed)  (size)   Mode  Cnt    Score   Error   Units
ByteBufIndexOfBenchmark.indexOf           true                  4             0       false     false       1       7  thrpt   16  120.986 ± 0.916  ops/us
ByteBufIndexOfBenchmark.indexOf           true                  4             0       false     false       1      16  thrpt   16   77.810 ± 0.603  ops/us
ByteBufIndexOfBenchmark.indexOf           true                  4             0       false     false       1      23  thrpt   16   62.605 ± 0.513  ops/us
ByteBufIndexOfBenchmark.indexOf           true                  4             0       false     false       1      32  thrpt   16   64.717 ± 0.333  ops/us
ByteBufIndexOfBenchmark.indexOf           true                 11             0       false     false       1       7  thrpt   16   82.222 ± 3.057  ops/us
ByteBufIndexOfBenchmark.indexOf           true                 11             0       false     false       1      16  thrpt   16   74.278 ± 2.926  ops/us
ByteBufIndexOfBenchmark.indexOf           true                 11             0       false     false       1      23  thrpt   16   60.679 ± 0.561  ops/us
ByteBufIndexOfBenchmark.indexOf           true                 11             0       false     false       1      32  thrpt   16   60.645 ± 0.745  ops/us
```
OLD:
```
Benchmark                             (direct)  (logPermutations)  (needleByte)  (noUnsafe)  (pooled)  (seed)  (size)   Mode  Cnt   Score   Error   Units
ByteBufIndexOfBenchmark.indexOf           true                  4             0       false     false       1       7  thrpt   16  97.411 ± 0.703  ops/us
ByteBufIndexOfBenchmark.indexOf           true                  4             0       false     false       1      16  thrpt   16  76.275 ± 0.402  ops/us
ByteBufIndexOfBenchmark.indexOf           true                  4             0       false     false       1      23  thrpt   16  61.636 ± 0.280  ops/us
ByteBufIndexOfBenchmark.indexOf           true                  4             0       false     false       1      32  thrpt   16  48.430 ± 0.170  ops/us
ByteBufIndexOfBenchmark.indexOf           true                 11             0       false     false       1       7  thrpt   16  76.215 ± 0.844  ops/us
ByteBufIndexOfBenchmark.indexOf           true                 11             0       false     false       1      16  thrpt   16  60.042 ± 1.890  ops/us
ByteBufIndexOfBenchmark.indexOf           true                 11             0       false     false       1      23  thrpt   16  43.772 ± 0.693  ops/us
ByteBufIndexOfBenchmark.indexOf           true                 11             0       false     false       1      32  thrpt   16  29.927 ± 0.205  ops/us
```

I've tried to improve the case with `size 7` by manually loop-unrolling the checks and indeed I'm getting back a nice speedup.